Remove timeout from flakey test
We don't need to have a timeout in the test because if stop fails to
terminate the process (SIGTERM) it should kill it after 10 seconds. If
we see this code hanging we could re-add the test timeout but until then
it is only causing flakes without adding value to the coverage.

Signed-off-by: George Lestaris <glestaris@pivotal.io>