Use a larger basebox for some quota limits tests
There was one test in particular which was flakey. It attempts to check
that the quota has been limited to the user-layer. So the user should be
able to write files up to the size of the full quota (plus or minus a
little noise for log files and btrfs compression), rather than having
some of their quota taken up by the basebox.

This test was flakey because the basebox we were using (busybox) was
extremely small. The noise generated by log files and compression was
sometimes larger than the basebox itself. For this test we therefore
choose to use a larger basebox: busyboxplus.  Dockerhub seems to claim
that this image is 4MB:
https://hub.docker.com/r/cloudfoundry/busyboxplus/tags/

...which is plenty.

[#104885948]

Signed-off-by: Zhou Yu <zyu@pivotal.io>