Fix NetIn flake
This is an attempt to fix the NetIn flake. We didn't manage to reproduce
it and we cannot be sure that this change is actually fixing or reducing
the occurances of then NetIn flake. We changed the way it waits for the
netcat server to come up by replacing a sleep with an assertion that
waits for `netstat` to include 'LISTEN'. When this happens inside the
container where netcat was started, we can be confident that the server
is running and therefore a good time to run the client against it.

[#113642763]

Signed-off-by: Alberto Leal <aleal@pivotal.io>