Stop Skipping Disk limit test
We used to skip this test when using an external image plugin as the
test would fail when configured with grootfs. This is because grootfs
has a smaller diskspace overhead, so it requires us to write more blocks
to get the test to fail. Now that we're making grootfs the default, we
can up the limit in this test and always run it.

[#152470748]

Signed-off-by: Ed King <eking@pivotal.io>