Add cpu burst test
The test starts 2 containers: miner and extra, miner weighing 1/3 and
extra weighing 2/3 in cpu shares. Miner is doing cpu intensive
operations, while extra is idle.

There is a 'burstControl' function that monitors container metrics and
throttles bursting containers for some period of time. A container is
said to be bursting if it consumes more cpu time than its proportional weight
in shares. For example miner is allowed to burst to 100% for a period
of time, then it is throttled to its baseline of 33% for two periods and
when the penalty expires it is unthrottled. The whole process repeats.