Keep stdin open until process command can open it
Fixes a race where dadoo/execrunner closes stdin fifo before `cat` opens
it via the /dev/stdin symlink, causing a hang.

Note it is up to the end-user client to ensure this behaviour. The
garden code (client and server) simply keeps the fifo open until the
end-user client closes it. So we use a pipe here to make that happen.

[#173495708]

Co-authored-by: Georgi Sabev <georgethebeatle@gmail.com>