Reduce flakes for GraceTime tests
These tests seem to be flakey in a way that indicates the containers
are never destroyed while waiting for the GraceTime to expire. Since
it is based on the time we last checked for the container state, we
may have a noisy neighbor problem with parallel tests on small/burstable
CPUs in CI.

Updating our query interval to be 4x the gracetime threshold, to make it
more likely that we won't accidentally re-trigger gracetime refreshing.