Rework the logic for detecting an idle app's ability to spike CPU while another app is throttled
Previously it ensured the ratio of goodContainerCPU to badContainerCPU
was less than 0.5 (bad container using 2x the cpu of good container),
however usage for each was based on how much CPU time was used divided
by its entitlement time.

This would imply that goodContainer should use NO MORE than 50% of
its CPU allocation while spinning, while badContainer would use 100% of
its smaller allocation.

We're changing this to detect that the goodContainer is able to spike,
by measuring CPU for 5s prior to telling it to spin, then for 5s after
telling it to spin, and ensuring that CPU usage increased at least 10x